+++ src/App.js
@@ -2,6 +2,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import RegisterPage from "./pages/RegisterPage";
 import LoginPage from "./pages/LoginPage";
 import HomePage from "./pages/HomePage";
+import CartPage from "./pages/CartPage";
 import MainLayout from "./layouts/MainLayout";
 import ProductDetailPage from "./pages/ProductDetailPage";
 import * as paths from "./constants/urls";
@@ -19,6 +20,7 @@ function App() {
               path={paths.PRODUCT_DETAIL_PATH}
               element={<ProductDetailPage />}
             />
+            <Route path={paths.CART_PATH} element={<CartPage />} />
           </Route>
         </Routes>
       </BrowserRouter>
