+++ src/components/atoms/Button.jsx
@@ -5,7 +5,9 @@ const Button = ({ onClick, children, className }) => {
   /** children: 버튼 내부에 표시될 내용 */
   return (
     <button
-      className={className} /**className: 클래스 이름 */
+      className={
+        "bg-yellow-500 text-black px-4 py-2 rounded"
+      } /**className: 클래스 이름 */
       onClick={(e) => {
         e.preventDefault();
         onClick(); // onClick이 없으면 빈 함수가 호출됩니다.
