+++ src/pages/Cartpage.jsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import CartList from "../components/molecules/CartList";
 import { Suspense, useEffect, useState } from "react";
-import { getCart } from "../apis/cart";
+import { getCart } from "../services/cart";
 import React from "react";
 import { useQuery } from "react-query";
 import Loader from "../components/atoms/Loader";
