+++ src/services/cart.js
@@ -1,5 +1,13 @@
 import { instance } from "./index";
 
-export const addCart = (payload) => {
-  return instance.post("/carts/add", payload);
+export const addCart = (items) => {
+  return instance.post("/carts/add", items);
+};
+
+export const getCart = () => {
+  return instance.get("/carts");
+};
+
+export const updateCart = (items) => {
+  return instance.post("/carts/update", items);
 };
